Address PR review: name-fallback, runHttp cause, thread-repair extraction
- errors.ts: isAbortError's cross-module name fallback now also matches
  "GuardExceededError" — a guard trip crossing the subprocess shim (where
  instanceof fails) is otherwise misclassified as a normal error and converted
  to a Failure instead of propagating. (+ a locking test.)
- http.ts runHttp: when the abort rejection carries no structured cause (a
  bare DOMException with no reason), fall back to reading the cause off the
  active runtime abort signal so a guardTrip survives the fetch boundary.
  (In current Node, fetch rejects with signal.reason, so the primary path
  already carries it; this hardens the no-reason edge.)
- errors.test.ts: the GuardExceededError subclass test now passes an explicit
  guardId and asserts it round-trips on the cause, keeping the owned-guard-id
  contract visible.
- Extract markThreadCancelled + needsThreadRepair out of the oversized
  prompt.ts into lib/runtime/threadRepair.ts (with threadRepair.test.ts);
  prompt.ts imports them. Pure move, no behavior change.

All green: make, full unit suite (6135), guards fixtures (25), lint:structure.
