+++ packages/agency-lang/lib/runtime/errors.ts
@@ -200,6 +200,7 @@ export function isAbortError(error: unknown): boolean {
     return (
       error.name === "AgencyAbort" ||
       error.name === "AgencyCancelledError" ||
+      error.name === "GuardExceededError" ||
       error.name === "AbortError"
     );
   }
