+++ packages/agency-lang/lib/stdlib/http.ts
@@ -88,12 +88,23 @@ export async function runHttp<T>(fn: () => Promise<T>, url: string): Promise<T> 
     return await fn();
   } catch (e) {
     if (isAbortError(e)) {
-      // Carry the structured cause off the rejection (it IS the signal's
-      // reason). A guard trip that aborted this fetch surfaces as an
-      // AgencyCancelledError whose guardTrip cause must survive so the
-      // owning guard's `try` converts it instead of letting a bare cancel
-      // escape. Falls back to a plain cancel for an external/DOMException abort.
-      throw new AgencyCancelledError(`fetch ${url} cancelled`, readCause(e));
+      // Carry the structured cause so a guard trip that aborted this fetch
+      // surfaces as an AgencyCancelledError whose guardTrip cause survives —
+      // the owning guard's `try` then converts it instead of letting a bare
+      // cancel escape. In current Node, `fetch` rejects with `signal.reason`
+      // directly, so the cause is on `e`. But an abort delivered as a bare
+      // DOMException (no reason) wouldn't carry it, so fall back to reading
+      // the cause off the active runtime abort signal.
+      let cause = readCause(e);
+      if (cause === undefined) {
+        try {
+          const { ctx, stack } = getRuntimeContext();
+          cause = readCause(ctx.getAbortSignal(stack));
+        } catch {
+          /* not inside an execution frame — no signal cause to recover */
+        }
+      }
+      throw new AgencyCancelledError(`fetch ${url} cancelled`, cause);
     }
     throw e;
   }
